fix: accept URL and DSN as connection string
- Fix isPostgresConnectionString to accept both URL and DSN
  formats, as promised in ReplaceDatabaseInConnectionString.
  Empty connection strings are prohibited for defensive
  programming;
- Putting migration runners' ordering functions into a
  separate file to simplify reading;
- Minor godoc and code-style changes.
